StatNLP PyTorch V0.6  Minimal Span Tree Example (fix bugs in compiler)

+++ hypergraph/NetworkConfig.py
@@ -2,7 +2,7 @@ import torch
 import numpy as np
 
 class NetworkConfig:
-    DEFAULT_CAPACITY_NETWORK = np.asarray([1000,1000, 1000, 1000, 1000])
+    DEFAULT_CAPACITY_NETWORK = np.asarray([1000,1000, 1000, 1000, 1000], dtype=int)
     NUM_THREADS = 1
     GPU_ID = -1
     DEVICE = torch.device("cpu")  #device = torch.device("cuda:" + args.gpuid)
